clean up dialog exports
